Move quiz tests from utils to test utils
