+++ ddd-koans/util/src/main/kotlin/Utils.kt
@@ -1,42 +1,1 @@
-import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatCode
-import org.junit.jupiter.api.Test
-
 fun Int.isPowerOfTwo() = this.and(this - 1) == 0
-
-abstract class QuizTest<T>(private val expectedAnswer: T, private val answerProvider: () -> T) {
-
-    @Test
-    fun testIfTheAnswerIsCorrect() {
-        assertThatCode {
-            answerProvider()
-        }
-            .`as`("Question method that returns answer should not throw exception.")
-            .doesNotThrowAnyException()
-
-        assertThat(clearAnswer(answerProvider()))
-            .`as`("Answer(s) should be as expected.")
-            .withFailMessage("Answer(s) should be as expected.")
-            .isEqualTo(clearAnswer(expectedAnswer))
-    }
-
-    protected abstract fun clearAnswer(answers: T): T
-}
-
-abstract class SingleQuizTest(expectedAnswer: String, answerProvider: () -> String) :
-    QuizTest<String>(expectedAnswer, answerProvider) {
-    override fun clearAnswer(answers: String) =
-        answers
-            .toLowerCase()
-            .trim()
-}
-
-abstract class MultiQuizTest(expectedAnswers: Set<String>, answerProvider: () -> Set<String>) :
-    QuizTest<Set<String>>(expectedAnswers, answerProvider) {
-
-    override fun clearAnswer(answers: Set<String>): Set<String> =
-        answers
-            .map(String::toLowerCase)
-            .map(String::trim)
-            .toSet()
-}
